app/requirements.txt
pyftpdlib
django
python-telegram-bot
djangorestframework
pyyaml
